requirements.txt
click>=8.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
pytesseract>=0.3.10
